requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
httpx==0.27.2
python-dotenv==1.0.1
PyGithub==2.5.0
tenacity==9.0.0
orjson==3.10.7
GitPython==3.1.43
pytest==8.3.3
playwright==1.48.0
pytest-playwright==0.5.2

